Allow contribution after end of ROSCA (#40)
* allow contributions after end of rosca and before collection

We need to allow participants to contribute after ROSCA has ended. Participants may occasionally miss a payment and should be allowed to contribute even after ROSCA has ended to catch up on their payments.
As a reminder, it is the foreperson's responsibility to settle the payments post-ROSCA as they can extract any surplus funds in the contract.
Allowing contributions post-ROSCA cannot cause a problem because the contract is self-destructed once the foreperson retrieves the funds.