Introduce Interim Fees Program (#50)
This is a stepping stone in getting to a perfect fee program.
Currently we calculate fees during contribution. However, this will result in fees being places also on over-contributions from participants, which we want to avoid.
In order to be able to extract only the required fees, we need to move to the contributions/winnings model Shine and I spoke about on slack.

This PR is therefore a stepping stone on the way to perfect fees. Therefore, the tests are also not perfect, they will be improved with the new accounting system.