Introduce an Escape Hatch to the ROSCA contract (#51)
Only WeTrust (aka the Escape Hatch Enabler) can enable the Escape Hatch
Once the EH is enabled, only the foreperson may activate it.
Once the EH is activated, conribute() and withdraw() throw, and emergencyWithdrawal() becomes available to be called, only by the foreperson. It selfdestructs the contract, sending all funds to the foreperson, to be dispersed (offline) to all participants.